% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{add_owm_tiles}
\alias{add_owm_tiles}
\title{Add owm tiles to leaflet map.}
\usage{
add_owm_tiles(map, layer_name = "temp", ...)
}
\arguments{
\item{map}{leaflet map object}

\item{layer_name}{owm layer name,
see \code{\link{owm_layers}}}

\item{...}{optional parameters passed to
\code{\link[leaflet]{addTiles}}}
}
\value{
updated map object
}
\description{
Add owm tiles to leaflet map.
}
\examples{
\dontrun{
   leaflet() \%>\% add_owm_tiles() \%>\%
      addMarkers(data = quakes[1:20, ])
}

}
