% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{current_as_tibble}
\alias{forecast_as_tibble}
\title{Parse owm data to tibble object}
\usage{
as_tibble(resp, simplify = TRUE, ...)

current_as_tibble(resp, simplify = TRUE, ...)

forecast_as_tibble(resp, simplify = TRUE, ...)
}
\arguments{
\item{resp}{(list) Result returned from \code{\link{get_current}}
or \code{\link{get_forecast}}}

\item{simplify}{(boolean) Should "extra" columns like ids and icons be comitted from the dataframe output?}

\item{...}{Further arguments.}
}
\value{
A dataframe containing weather-relevant information.
}
\description{
Parse owm data to tibble object
}
