% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up-cbind_weather.R
\name{cbind_weather}
\alias{cbind_weather}
\title{Flatten weather column in data frame.}
\usage{
cbind_weather(data)
}
\arguments{
\item{data}{data frame containing weather column}
}
\value{
data frame with flattened weather (data)
}
\description{
Flatten weather column in data frame.
}
\examples{
\dontrun{
   get_forecast("Kassel") \%>\% cbind_weather()
}
}
