% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-test-data.R
\name{fetch_test_data}
\alias{fetch_test_data}
\title{helper function to fetch test data}
\usage{
fetch_test_data(run_test = TRUE)
}
\arguments{
\item{run_tests}{set to true in order to run tests after data
was fetched succesfully}
}
\description{
helper function to update test data
before running tests, don't forget to set api key
via \code{owmr_setup} before trying to fetch data
}

