% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-multiple.R
\name{find_cities_by_geo_point}
\alias{find_cities_by_geo_point}
\title{Find cities by geo point.}
\usage{
find_cities_by_geo_point(lat, lon, cnt = 3, ...)
}
\arguments{
\item{lat}{latitude of geo point}

\item{lon}{longitude of geo point}

\item{cnt}{number of cities}

\item{...}{see owm api documentation}
}
\value{
list containing data frame with weather data
}
\description{
Get current weather data for a number of cities
around given geo point.
}
\examples{
\dontrun{
  find_cities_by_geo_point(lat = 51.50853, lon = -0.12574, cnt = 5)
}
}
\seealso{
\code{\link{find_city}}
}
