% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{find_city}
\alias{find_city}
\title{find city by name or (lat, lon)}
\usage{
find_city(city = NA, ...)
}
\arguments{
\item{city}{city name (and country code)}

\item{...}{see owm api}
}
\value{
weather list for matches
}
\description{
either search for city by name or fetch weather
for a number of cities around geo point
}
\examples{
\dontrun{
   find_city("London,UK")
   find_city(lat = 51.50853, lon = -0.12574, cnt = 5)
}
}

