% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{find_stations_by_geo_point}
\alias{find_stations_by_geo_point}
\title{Find stations by geo point.}
\usage{
find_stations_by_geo_point(lat, lon, cnt = 10, ...)
}
\arguments{
\item{lat}{latitude of geo point}

\item{lon}{longitude of geo point}

\item{cnt}{number of stations}

\item{...}{see owm api documentation}
}
\value{
data frame
}
\description{
Get weather data from a number of stations
around given geo point.
}
\examples{
\dontrun{
   # get weather data from 7 stations
   find_stations_by_geo_point(lat = 51.31667, lon = 9.5, cnt = 7)
}
}
