% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up-flatten.R
\name{flatten}
\alias{flatten}
\title{Flatten list.}
\usage{
flatten(data)
}
\arguments{
\item{data}{list returned from owm}
}
\value{
flattened list
}
\description{
Flatten list.
}
\examples{
\dontrun{
   get_current("Rio de Janeiro") \%>\% flatten()
   get_current("Rio de Janeiro") \%>\% flatten() \%>\%
      tidy_up_()
}
}
