% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_.R
\name{flatten_}
\alias{flatten_}
\title{Flatten nested data frames.}
\usage{
flatten_(data, sep = ".")
}
\arguments{
\item{data}{data frame}

\item{sep}{separator used to combine nested (list) column names with column
names of nested data frames}
}
\value{
unnested/flattened data frame
}
\description{
Just wraps \code{\link[tidyr]{unnest}} (package \pkg{tidyr}
setting paremeter \code{sep = "."} by default.
}
\seealso{
\code{\link[tidyr]{unnest}}
}

