% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-multiple.R
\name{get_current_for_group}
\alias{get_current_for_group}
\title{Title}
\usage{
get_current_for_group(city_ids, ...)
}
\arguments{
\item{city_ids}{vector containing city ids,
see \code{\link{owm_cities}} dataset for ids}

\item{...}{see owm api (units, ...)}
}
\value{
list containing data frame with weather data of cities
}
\description{
Title
}
\examples{
\dontrun{
   city_ids = c(2831088, 2847639, 2873291)
   result <-get_current_for_group(city_ids)
   weather_frame <- result$list
}
}

