% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_current_from_station}
\alias{get_current_from_station}
\title{get current weather data from given station}
\usage{
get_current_from_station(station_id, ...)
}
\arguments{
\item{station_id}{station id}

\item{...}{see owm api}
}
\value{
list
}
\description{
get current weather data from given station
}

