% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get 3h forecast.}
\usage{
get_forecast(city = NA, ...)
}
\arguments{
\item{city}{city name or id}

\item{...}{see owm api documentation}
}
\value{
list
}
\description{
Get 3h forecast.
}
\examples{
\dontrun{
   result <- get_forecast("Kassel", units = "metric")
   names(result)
}
}

