% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-daily.R
\name{get_forecast_daily}
\alias{get_forecast_daily}
\title{Get daily forcast up to 16 days.}
\usage{
get_forecast_daily(city = NA, ...)
}
\arguments{
\item{city}{city name or id}

\item{...}{see owm api documentation}
}
\value{
list
}
\description{
Get daily forcast up to 16 days.
}
\examples{
\dontrun{
   # 9 day forecast
   result <- get_forecast_daily("London", cnt = 9)
   forecast_frame <- result$list
}
}

