% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_.R
\name{knock_down}
\alias{knock_down}
\title{Flatten nested data frames.}
\usage{
knock_down(data, sep = ".")
}
\arguments{
\item{data}{data frame containing nested data frames}

\item{sep}{separator used to combine flattened column names
with column names of nested data frames}
}
\value{
flattened data frame
}
\description{
Wraps \code{\link[tidyr]{unnest}} from package \pkg{tidyr}
setting parameter \code{sep = "."} by default.
}
\examples{
\dontrun{
   find_city("Malaga")$list \%>\% knock_down()
}
}
\seealso{
\code{\link[tidyr]{unnest}}
}

