% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_weather}
\alias{merge_weather}
\title{Merge weather.}
\usage{
merge_weather(x, prefix = "weather.")
}
\arguments{
\item{x}{main data frame returned from owm}

\item{prefix}{prefix added to names of flattened weather data frame}
}
\value{
main data frame with weather data frame appended
}
\description{
Flatten weather column (list of data frames) and append it to main (top level) data frame.
}

