% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owm_cities}
\alias{owm_cities}
\title{owm city list containing id and coords of cities.}
\format{A data frame with 74071 rows and 4 variables:
\describe{
  \item{id}{city id}
  \item{nm}{city name}
  \item{lat}{latitude}
  \item{lon}{longitude}
  \item{countryCode}{two letter country code}
}}
\source{
\url{http://openweathermap.org/help/city_list.txt}
}
\usage{
owm_cities
}
\description{
A dataset containing city ids which can be used in queries
}
\keyword{datasets}

