% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owm_layers}
\alias{owm_layers}
\title{List of available owm weather map layers.}
\format{An object of class \code{list} of length 16.}
\usage{
owm_layers
}
\description{
List of available owm weather map layers.
}
\seealso{
\url{https://openweathermap.org/api/weathermaps}
}
\keyword{datasets}
