% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{owmr}
\alias{owmr}
\title{owmr - An R interface to OpenWeatherMap}
\description{
owmr - An R interface to OpenWeatherMap
}
\examples{
\dontrun{
   get_current("Kassel", units = "metric")
   get_forecast("London", cnt = 7)
   # ...
}

}

