% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-to-tibble.R
\name{owmr_as_tibble}
\alias{owmr_as_tibble}
\alias{owmr_as_tibble.owmr_weather}
\alias{owmr_as_tibble.default}
\alias{owmr_as_tibble.owmr_forecast_daily}
\title{Parse owmr response to tibble.}
\usage{
owmr_as_tibble(resp, simplify = TRUE)

\method{owmr_as_tibble}{owmr_weather}(resp, simplify = TRUE)

\method{owmr_as_tibble}{default}(resp, simplify = TRUE)

\method{owmr_as_tibble}{owmr_forecast_daily}(resp, simplify = TRUE)
}
\arguments{
\item{resp}{response object returned from functions like
\code{\link{get_current}} or \code{\link{get_forecast}}}

\item{simplify}{return tibble only?}
}
\value{
list containing tibble or tibble only (\code{simplify = TRUE})
}
\description{
Parse owmr response to tibble.
}
