% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{"\%$\%"}
\alias{render}
\title{Render template.}
\usage{
render(template, data)
}
\arguments{
\item{template}{template}

\item{data}{data frame where column names correspond to
variables names in template}
}
\value{
rendered template
}
\description{
Vectorizes \code{\link[whiskers]{whisker.render}}.
}
\examples{
vars <- data.frame(a = 1:3, b = 23:21)
"a = {{a}} and b = {{b}}" \%$\% vars
}
\seealso{
\code{\link[whiskers]{whisker.render}}
}

