% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-list.R
\name{search_city_list}
\alias{search_city_list}
\title{search for given city in owm's city list}
\usage{
search_city_list(city, country_code = "")
}
\arguments{
\item{city}{city name}

\item{country_code}{two letter country code (AU, DE, ...)}
}
\value{
data frame with matches
}
\description{
see also \code{owm_cities} dataset
}
\examples{
# TODO
}

