% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up.R
\name{tidy_up}
\alias{tidy_up}
\title{Tidy up owm data.}
\usage{
tidy_up(data, use_underscore = TRUE, remove_prefix = c("main", "sys"))
}
\arguments{
\item{data}{response (list) returned from owm}

\item{use_underscore}{substitute dots in column names with an underscore}

\item{remove_prefix}{prefices to be removed for shorter column names
(\code{remove_prefix = NULL}) will keep all prefices}
}
\value{
response with 'clean' data frame
}
\description{
Tidy up owm data.
}
\examples{
\dontrun{
   result <- find_city("Malaga") \%>\% tidy_up()
   # keep dots in column names
   result <- find_city("Malaga") \%>\% tidy_up(use_underscore = FALSE)
   # keep all prefices
   result <- find_city("Malaga") \%>\% tidy_up(remove_prefix = NULL)
}
}

