% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up-use_underscore.R
\name{use_underscore}
\alias{use_underscore}
\title{Substitute dots in column names with underscores.}
\usage{
use_underscore(data)
}
\arguments{
\item{data}{data frame}
}
\value{
data frame with updated column names
}
\description{
Substitute dots in column names with underscores.
}
\examples{
names(airquality)
use_underscore(airquality) \%>\% names()
}
