% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-popup.R
\name{wrap}
\alias{wrap}
\title{Wrap html tag around value.}
\usage{
wrap(tag, value)
}
\arguments{
\item{tag}{tag like \code{i}, \code{strong} or whatever}

\item{value}{value to be wrapped around}
}
\value{
wrapped value
}
\description{
Wrap html tag around value.
}
\examples{
x <- get_current("London", units = "metric")
# create a simple popup to show up on a (leaflet) map
popup <- wrap("strong", x$main$temp)

}

