% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{current_as_tibble}
\alias{forecast_as_tibble}
\title{Parse owm data to tibble object}
\usage{
current_as_tibble(data)

forecast_as_tibble(data)
}
\arguments{
\item{data}{result returned from \code{\link{get_current}}
or \code{\link{get_forecast}}}
}
\description{
Parse owm data to tibble object
}
