% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_test_data.R
\name{fetch_test_data}
\alias{fetch_test_data}
\title{Helper function to fetch test data.}
\usage{
fetch_test_data(run_tests = TRUE)
}
\arguments{
\item{run_tests}{set \code{run_tests = TRUE} in order
to run tests after data was fetched succesfully}
}
\description{
Updates test data before running tests.
}

