% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up-flatten_weather.R
\name{flatten_weather}
\alias{flatten_weather}
\title{Parse weather column to (single) data frame.}
\usage{
flatten_weather(x)
}
\arguments{
\item{x}{weather column (NOT name)}
}
\value{
data frame
}
\description{
Parse weather column to (single) data frame.
}
\examples{
\dontrun{
   result <- get_forecast("Kassel", units = "metric")$list
   weather <- flatten_weather(result$weather)
   weather$description \%>\% print()
}
}
