% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{get_current}
\alias{get_current}
\title{get current weather for given city}
\usage{
get_current(city = NA, ...)
}
\arguments{
\item{city}{city name or id}

\item{...}{see owm api for optional parameters}
}
\value{
list
}
\description{
get current weather for given city
}
\examples{
\dontrun{
   get_current("London", units = "metric")
}
}

