% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{get_current}
\alias{get_current}
\title{Get current weather data for given city.}
\usage{
get_current(city = NA, ...)
}
\arguments{
\item{city}{city name or id}

\item{...}{see owm api documentation, you can also skip parameter
\code{city} and pass \code{lat} (latitude) and \code{lon} (longitude)
or \code{zip} (zip code) instead}
}
\value{
list
}
\description{
Get current weather data for given city.
}
\examples{
\dontrun{
   get_current("London", units = "metric")
   get_current(2643741, lang = "DE")
   get_current(lon = -0.09184, lat = 51.51279)
   get_current(zip = "94040,US")
}
}
