% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{get forecast}
\usage{
get_forecast(city, ...)
}
\arguments{
\item{city}{city name}

\item{...}{see owm api for optional parameters}
}
\value{
list
}
\description{
get forecast
}
\examples{
\dontrun{
   get_forecast("Kassel", units = "metric")
}
}

