% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-daily.R
\name{get_forecast_daily}
\alias{get_forecast_daily}
\title{get daily forcast up to 16 days}
\usage{
get_forecast_daily(loc = NA, ...)
}
\arguments{
\item{loc}{city name}

\item{...}{see owm api parameters}
}
\value{
list
}
\description{
get daily forcast up to 16 days
}
\examples{
\dontrun{
   # 9 days forecast
   get_forecast_daily("London", cnt = 9)
}
}

