% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{owmr}
\alias{owmr}
\title{owmr - An R interface to OpenWeatherMap}
\description{
owmr - An R interface to OpenWeatherMap
}
\examples{
\dontrun{
   # first of all you have to set up your api key
   owmr_settings("your_api_key")

   # get current weather for "Kassel" with temperatures in °C
   get_current("Kassel", units = "metric")

   # get 7 day forcast
   get_forecast("London", cnt = 7)

   # ...
}

}

