% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{owmr_settings}
\alias{owmr_settings}
\title{owmr settings.}
\usage{
owmr_settings(api_key)
}
\arguments{
\item{api_key}{owm api key}
}
\description{
Set api key.
}
\examples{
\dontrun{
   owmr_settings(api_key = "your-api-key")
}
}
