% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_result}
\alias{parse_result}
\title{Apply functions to columns of result.}
\usage{
parse_result(x, functions_)
}
\arguments{
\item{x}{result data frame}

\item{functions_}{named where keys correspond to column names}
}
\value{
parsed result data frame
}
\description{
Apply functions to columns of result.
}
\examples{
\dontrun{
   parse_dt <- function(x){as.POSIXct(x, origin = "1970-01-01")}
   forecast <- get_forecast("Kassel")$list
   forecast \%<>\% parse_result(list(dt = parse_dt))
}
}

