% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-popup.R
\name{popup}
\alias{popup}
\title{Popup creator.}
\usage{
popup(x, ...)
}
\arguments{
\item{x}{data, usually a result data frame as returned by
\code{\link{get_forecast}} or
\code{\link{get_current_for_group}}}

\item{...}{see \code{\link{paste0}}}
}
\value{
pasted objects
}
\description{
Calls \code{\link{paste0}} in an environment constructed from passed
data.
}
\examples{
\dontrun{
   df <- get_forcast("London", units = "metric", cnt = 7)$list
   popup(df, dt_txt, "</br>", main.temp)
}
}

