% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{\%$$\%}
\alias{\%$$\%}
\alias{render}
\title{Render operator.}
\usage{
template \%$$\% data
}
\arguments{
\item{template}{template}

\item{data}{data frame where column names correspond to
variables names in template}
}
\value{
rendered template
}
\description{
Vectorizes function \code{\link[whisker]{whisker.render}}. \cr
\cr
NOTE: Because \pkg{whisker} does not support variable names
inlcuding dots, a \emph{dot} in column names is replaced by an \emph{underscore}.
Therefore, you must use an underscore in the template text for
varibales including dots.
}
\examples{
vars <- data.frame(a = 1:3, b = 23:21)
"a = {{a}} and b = {{b}}" \%$$\% vars
}
\seealso{
\code{\link[whisker]{whisker.render}}
}
