% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_list.R
\name{search_city_list}
\alias{search_city_list}
\title{Look up coordinates and city id in owm's city list.}
\usage{
search_city_list(city, country_code = "")
}
\arguments{
\item{city}{city name (regex)}

\item{country_code}{two letter country code (AU, DE, ...),
use \code{country_code = ""} as wildcard}
}
\value{
data frame with matches
}
\description{
search \code{\link{owm_cities}} dataset by city name
and country code
}
\examples{
search_city_list("London", "GB")
search_city_list("London")
search_city_list("Lond")
}
\seealso{
\code{\link{owm_cities}} dataset
}
