% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up.R
\name{tidy_up}
\alias{tidy_up}
\title{Tidy up owm data.}
\usage{
tidy_up(data, ...)
}
\arguments{
\item{data}{result returned from owm
containing data frame in \code{data$list}}

\item{...}{see \code{\link{tidy_up_}}}
}
\value{
data with updated data frame (\code{data$list})
}
\description{
Calls \code{\link{tidy_up_}} passing \code{data$list}
   as \code{data} argument.
}
\examples{
\dontrun{
   get_forecast("London") \%>\% tidy_up()
}
}
\seealso{
\code{\link{tidy_up_}}
}
